requirements.txt
Django~=2.2.13
django-admin-rangefilter
django-widget-tweaks